+++ src/app/past-winners/_data/past-winners/pastWinners.ts
@@ -1,0 +1,10 @@
+import { pastWinners2021 } from './pastWinners2021';
+import { pastWinners2022 } from './pastWinners2022';
+import { pastWinners2023 } from './pastWinners2023';
+import { pastWinners2024 } from './pastWinners2024';
+export const pastWinners = [
+  pastWinners2024,
+  pastWinners2023,
+  pastWinners2022,
+  pastWinners2021,
+];
