+++ src/app/past-winners/page.tsx
@@ -1,10 +1,7 @@
 'use client';
 
 import React, { useState } from 'react';
-import { pastWinners2021 } from './_data/past-winners/pastWinners2021';
-import { pastWinners2022 } from './_data/past-winners/pastWinners2022';
-import { pastWinners2023 } from './_data/past-winners/pastWinners2023';
-import { pastWinners2024 } from './_data/past-winners/pastWinners2024';
+import { pastWinners } from './_data/past-winners/pastWinners';
 import { H1 } from '@/components/layout/Typography';
 import { SectionLayout } from '@/components/layout/SectionLayout';
 import { motion } from 'motion/react';
@@ -14,12 +11,6 @@ import { WinnersGroup } from '@/app/past-winners/_components/WinnersGroup';
 import { WinnerCard } from '@/app/past-winners/_components/WinnerCard';
 
 export default function PastWinnersPage() {
-  const pastWinners = [
-    pastWinners2024,
-    pastWinners2023,
-    pastWinners2022,
-    pastWinners2021,
-  ];
   const [selectedYear, setSelectedYear] = useState<number[]>([]); // Multiple years can be selected
   const [selectedPlace, setSelectedPlace] = useState<string[]>([]); // Multiple places can be selected
