+++ src/styles/globals.css
@@ -2,8 +2,10 @@
 @tailwind components;
 @tailwind utilities;
 
-* {
-  @apply text-white;
+@layer base {
+  body {
+    @apply text-white;
+  }
 }
 
 /* Navigation */
